Removes the #.png tacked to the end of google images.
I hide gifs by default in hipchat because, frankly, it's one of the
only tools I have to control the noise caused by people using the google
image lookup. Tacking `#.png` to the end of the url generated by this script
is enough to defeat Hipchat's gif hiding.

I don't know what problem this was trying to solve, but it seem superflous.